requirements.txt
streamlit==1.45.1
pandas
pymongo
langchain-openai
langgraph
google-api-python-client
python-dateutil
openai 
langchain 
langchain-community
langgraph
tavily-python
langgraph-checkpoint-sqlite
tabulate
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
streamlit-mic-recorder
ics